perf(ng-packagr): use native sass-embedded compiler daemon
Replaces the Piscina-based worker thread pool executing pure-JS Dart Sass
(`sass` compiled via `dart2js`) with a native Dart Sass asynchronous compiler
daemon (`sass-embedded`).

Previously, multi-threaded worker pools with synchronous compilation were used
to bypass performance overheads inherent to pure-JS asynchronous Dart Sass.
However, this incurred significant V8 heap and isolate duplication across worker
threads, as well as thread contention and blocking due to `Atomics.wait()`
synchronization loops during custom URL and module rebasing.

With `sass-embedded`, Sass compilation is delegated to a long-lived native Dart
AOT binary communicating asynchronously via Protocol Buffers over stdio pipes.
This eliminates duplicate V8 isolates, replaces synchronous `Atomics.wait()`
with event-driven asynchronous message handling, and reduces memory pressure
on the Node.js process.

The compiler infrastructure is structured around a shared `SassServiceImplementation`
interface with two isolated implementations dynamically loaded on demand:
- `SassAsyncCompilerImplementation`: Manages the native `sass-embedded` daemon lifecycle
  and asynchronous URL rebasing importers (`AsyncModuleUrlRebasingImporter`).
- `SassWorkerImplementation`: Retained as an isolated worker pool fallback.

Environment Variable & Fallbacks:
- `NG_BUILD_SASS_WORKER=1` (or `true`): Forces fallback to the pure-JS
  `SassWorkerImplementation` using worker threads.
- In restricted environments where native platform binaries cannot be spawned
  (such as WebContainers detected via `process.versions.webcontainer`), ng-packagr
  automatically falls back to `SassWorkerImplementation`.
