+++ package.json
@@ -50,7 +50,8 @@
     "rolldown": "^1.2.0",
     "rolldown-plugin-dts": "^0.28.0",
     "rxjs": "^7.8.1",
-    "sass": "^1.81.0",
+    "sass": "^1.103.1",
+    "sass-embedded": "^1.103.1",
     "tinyglobby": "^0.2.12"
   },
   "peerDependencies": {
